patch: add blank handlers for iq and pressense
