+++ chat/pkg/client/utils.go
@@ -1,6 +1,11 @@
 package client
 
-import "log"
+import (
+	"encoding/json"
+	"log"
+
+	"gosrc.io/xmpp/stanza"
+)
 
 // Get user Xmpp Auth token from ejabberd server
 func getUserXMPPAuthToken(jid string) (string, error) {
@@ -10,4 +15,16 @@ func getUserXMPPAuthToken(jid string) (string, error) {
 
 func errorHandler(err error) {
 	log.Println(err)
+}
+
+func compileMessage(msg stanza.Message) []byte {
+	m, err := json.Marshal(map[string]interface{}{
+		"from": msg.From,
+		"to":   msg.To,
+		"body": msg.Body,
+	})
+	if err != nil {
+		return nil
+	}
+	return m
 }
